backend/requirements.txt
# Testing
pytest==7.4.3
pytest-flask==1.2.0

# API Documentation
flask-swagger-ui==4.11.1

# Web Framework and Extensions
Flask==2.3.3
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-CORS==6.0.1
Flask-Migrate==4.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
requests==2.31.0
cloudinary==1.36.0
sendgrid==6.11.0
